pyproject.toml
[tool.poetry]
name = "dehefe assistant"
version = "0.1.0"
description = ""
authors = ["Anderson Rodrigues <eusouanderson@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pyttsx3 = "^2.98"
speechrecognition = "^3.14.0"
sounddevice = "^0.5.1"
pyaudio = "^0.2.14"
flask-socketio = "^5.5.1"
eventlet = "^0.38.2"
unicode = "^2.9"
unidecode = "^1.3.8"
fuzzywuzzy = "^0.18.0"
python-levenshtein = "^0.26.1"
psutil = "^6.1.1"
wmi = "^1.5.1"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
selenium = "^4.28.0"
webdriver-manager = "^4.0.2"
pytube = "^15.0.0"
pygame = "^2.6.1"
yt-dlp = "^2025.1.15"
python-vlc = "^3.0.21203"
googlesearch-python = "^1.3.0"
pytz = "^2024.2"
google-generativeai = "^0.8.5"
python-dotenv = "^1.1.1"
playsound = "1.2.2"
waitress = "^3.0.2"

[tool.poetry.scripts]
start = "poetry run python main.py"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.13"
pytest = "^8.4.0"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--cov=fastapi_zero --cov-report=term-missing -p no:warnings"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'poetry run python main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'